// src/app/model/project.ts
export class Project {

    projectId: number;
    applicationId: number;
    projectUid: string;
    projectName: string;
    startDate: Date;
    endDate: Date;
    releaseInfo: string;
    projectType: string;
    projectStatus: string;
    effort: number;
    tier2: string;
    primarySe: string;
    primaryDeveloper: string;
    primaryTester: string;
    manager: string;
    leadManager: string;


}